makefile updated

+++ src/include/stdlib.h
@@ -3,8 +3,8 @@
 
 #include <stddef.h>
 
-void* operator new (unsigned int size) throw() {  return 0; }
-void* operator new[] (unsigned int size) throw() {  return 0; }
+void* operator new (size_t size) throw() {  return 0; }
+void* operator new[] (size_t size) throw() {  return 0; }
 void operator delete (void * p) {  }
 void operator delete[] (void * p) {  }
